Remove leaf entry page operation for cell b-tree multi-value bucket v2 was added.

+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/OCheckpointEndRecord.java
@@ -54,7 +54,7 @@ public class OCheckpointEndRecord extends OAbstractWALRecord {
   }
 
   @Override
-  public byte getId() {
+  public int getId() {
     return WALRecordTypes.CHECKPOINT_END_RECORD;
   }
 }
