+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/po/cellbtree/singlevalue/v3/bucket/CellBTreeBucketSingleValueV3RemoveNonLeafEntryPO.java
@@ -60,7 +60,7 @@ public final class CellBTreeBucketSingleValueV3RemoveNonLeafEntryPO extends Page
   }
 
   @Override
-  public byte getId() {
+  public int getId() {
     return WALRecordTypes.CELL_BTREE_BUCKET_SINGLE_VALUE_V3_REMOVE_NON_LEAF_ENTRY_PO;
   }
