+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/po/cellbtree/singlevalue/v3/bucket/CellBTreeBucketSingleValueV3SetLeftSiblingPO.java
@@ -41,7 +41,7 @@ public final class CellBTreeBucketSingleValueV3SetLeftSiblingPO extends PageOper
   }
 
   @Override
-  public byte getId() {
+  public int getId() {
     return WALRecordTypes.CELL_BTREE_BUCKET_SINGLE_VALUE_V3_SET_LEFT_SIBLING_PO;
   }
